# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Muttley API
  description: >-
    API REST do Sistema Muttley para Gestão de Eventos Acadêmicos, Inscrições,
    Emissão e Validação de Certificados com Assinatura Visual e Gamificação por Medalhas.
    Desenvolvido para a FATEC Zona Leste.
  version: 0.1.0-alpha
  contact:
    name: Equipe Muttley - FATEC Zona Leste
    url: https://github.com/andrelamego/Backend-Muttley
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8083
    description: Servidor Local de Desenvolvimento
...
paths:
...
  /api/me:
    get:
      tags:
        - Meu Perfil
      summary: Obter perfil autenticado
      description: Retorna os dados da pessoa vinculada ao token JWT atual.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados retornados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PessoaMeResponse'
        '401':
          description: Não autenticado
        '404':
          description: Usuário não encontrado
  /api/me/certificados:
    get:
      tags:
        - Meu Perfil
      summary: Listar certificados do usuário
      description: Retorna todos os certificados emitidos para o usuário autenticado.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de certificados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CertificadoUsuarioResponse'
        '401':
          description: Não autenticado
  /api/me/medalhas:
    get:
      tags:
        - Meu Perfil
      summary: Listar medalhas do usuário
      description: Retorna as medalhas conquistadas pelo usuário autenticado.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de medalhas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedalhaUsuarioResponse'
        '401':
          description: Não autenticado
  /api/me/participacoes:
    get:
      tags:
        - Meu Perfil
      summary: Listar participações do usuário
      description: Retorna histórico de participações do usuário autenticado.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de participações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParticipacaoUsuarioResponse'
        '401':
          description: Não autenticado
...
  /api/participacoes:
    get:
      tags:
        - Participações
      summary: Listar participações permitidas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de participações
        '401':
          description: Não autenticado
    post:
      tags:
        - Participações
      summary: Criar nova participação
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoParticipacao'
      responses:
        '201':
          description: Participação criada
        '400':
          description: Dados inválidos
        '401':
          description: Não autenticado
  /api/participacoes/{id}:
    get:
      tags:
        - Participações
      summary: Buscar participação por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Participação encontrada
        '401':
          description: Não autenticado
        '404':
          description: Não encontrada
    put:
      tags:
        - Participações
      summary: Atualizar participação
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoParticipacao'
      responses:
        '200':
          description: Atualizada com sucesso
        '404':
          description: Não encontrada
    delete:
      tags:
        - Participações
      summary: Deletar participação
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deletada com sucesso
        '404':
          description: Não encontrada
...
  /api/admin/medalhas:
    get:
      tags:
        - Administração - Medalhas
      summary: Listar medalhas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de medalhas
    post:
      tags:
        - Administração - Medalhas
      summary: Criar medalha
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoMedalha'
      responses:
        '201':
          description: Criada com sucesso
  /api/admin/medalhas/{id}:
    get:
      tags:
        - Administração - Medalhas
      summary: Buscar medalha por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Encontrada
    put:
      tags:
        - Administração - Medalhas
      summary: Atualizar medalha
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoMedalha'
      responses:
        '200':
          description: Atualizada
    delete:
      tags:
        - Administração - Medalhas
      summary: Excluir medalha
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Excluída
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Insira o token JWT retornado no endpoint /api/auth/login.
  schemas:
...
    PessoaMeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        role:
          type: string
        aluno:
          type: boolean
        professor:
          type: boolean
        palestrante:
          type: boolean
        organizador:
          type: boolean
        colaborador:
          type: boolean
    CertificadoUsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        codigoValidacao:
          type: string
        dataEmissao:
          type: string
          format: date
        eventoTema:
          type: string
        eventoData:
          type: string
          format: date
    MedalhaUsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        tipo:
          type: string
        eventoTema:
          type: string
    ParticipacaoUsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        inscricao:
          type: string
        tipo:
          type: string
        presente:
          type: boolean
        eventoId:
          type: integer
          format: int64
        eventoTema:
          type: string
        eventoData:
          type: string
          format: date
...
    AtualizacaoParticipacao:
      type: object
      required:
        - pessoaId
        - eventoId
      properties:
        id:
          type: integer
          format: int64
        inscricao:
          type: string
        tipo:
          type: string
          example: ALUNO
        pessoaId:
          type: integer
          format: int64
        eventoId:
          type: integer
          format: int64
...
    AtualizacaoMedalha:
      type: object
      required:
        - nome
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          example: Participação Ouro
        descricao:
          type: string
        caminhoImagem:
          type: string
...